% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{print.binary_segmentation_result}
\alias{print.binary_segmentation_result}
\title{Print an object of class `binary_segmentation_result`.}
\usage{
\method{print}{binary_segmentation_result}(x, ...)
}
\arguments{
\item{x}{An object of class `binary_segmentation_result`.}

\item{...}{Not used.}
}
\description{
Print an object of class `binary_segmentation_result`.
}
